requirements.txt
numpy~=1.22.3
pillow~=9.2.0
omegaconf~=2.1.1
einops~=0.3.0
yaml~=0.2.5
pyyaml~=6.0
piexif~=1.1.3
transformers~=4.19.2
tqdm~=4.64.1
torchvision~=0.12.0
opencv-python~=4.6.0.66
pynvml~=11.4.1
scikit-image~=0.19.2
scipy~=1.7.3
retry~=0.9.2
streamlit~=1.13.0
albumentations~=0.4.3
clip~=1.0
kornia~=0.6.0
ftfy~=6.1.1
regex~=2019.11.1
setuptools~=59.8.0
midas~=2.1.0
timm~=0.6.7
zlib~=1.2.12
future~=0.18.2
pygments~=2.13.0
facexlib~=0.2.5
gfpgan~=1.3.8
basicsr~=1.4.2
realesrgan~=0.3.0
wandb~=0.13.3
matplotlib~=3.6.0
accelerate~=0.12.0
jsonmerge~=1.8.0
torchdiffeq~=0.2.3
requests~=2.25.1
packaging~=21.3
diffusers~=0.2.4
pandas~=1.4.3
multipart~=0.2.4
python-multipart~=0.0.5
flask~=2.1.1
eventlet~=0.33.0
werkzeug~=2.1.1
asgiref~=3.5.2
barfi~=0.7.0
